pyproject.toml
[tool.poetry]
name = "shark"
version = "0.1.0"
authors = ["Michael P. Jung <michael.jung@terreon.de>"]
license = "MIT OR Apache-2.0"
readme = "README.md"
description = "Django based billing application"
repository = "https://github.com/bikeshedder/shark"
homepage = "https://github.com/bikeshedder/shark"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5"
django-grappelli = "^4"
dinbrief = "^0.4.0"
django-taggit = "^5"
django-localflavor = "^4"
djangorestframework = "^3.12.4"
django-composite-field = "^2"
django-countries = "^7.2.1"
django-storages = "^1.14"
python-magic = "^0.4.24"
Wand = "^0.6.7"
PyPDF3 = "^1.0.5"
django-environ = "^0.11"
gunicorn = "^22"
boto3 = "^1.34"
psycopg = "^3.1.18"
django-tailwind = { extras = ["reload"], version = "^3.8.0" }
typst = "^0.11.0"
django-htmx = "^1.17.3"
heroicons = { extras = ["django"], version = "*" }
supercollect = "^1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4"
pre-commit = "^3.6.2"

[tool.ruff.lint]
# https://docs.astral.sh/ruff/rules/
select = ["E", "W", "F", "DJ", "I", "N"]

[tool.ruff.lint.per-file-ignores]
# Ignore `E501` (Line too long) in all migration files
"**/migrations/*" = ["E501"]

[tool.ruff.format]
line-ending = "lf"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
